% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_dropbox.R
\name{activate_dropbox}
\alias{activate_dropbox}
\title{Activate Dropbox access token}
\usage{
activate_dropbox(
  Dropbox_App_Name,
  Dropbox_App_Key,
  Dropbox_App_Secret,
  BW_app_path
)
}
\arguments{
\item{Dropbox_App_Name}{a character string, the name given to your Dropbox
App.}

\item{Dropbox_App_Key}{a character string, the Dropbox App Key. See details
for more information.}

\item{Dropbox_App_Secret}{a character string, the Dropbox App Secret. See
details for more information.}

\item{BW_app_path}{a character string, the path to your BIO-WELL Shiny App
folder.}
}
\value{
Saves a long-lived Dropbox token to your BIO-WELL survey Shiny App
folder.
}
\description{
Function to generate a long-lived Dropbox access token for app folder.
}
\details{
When your BIO-WELL survey Shiny App is hosted on the \href{https://posit.co/products/open-source/shinyserver/}{Shiny Server}, you will need to
save participant responses to a remote \href{https://www.dropbox.com/home}{Dropbox folder}.

This requires you to register for a free Dropbox account. To generate a
long-lived access token (over four hours) for your BIO-WELL survey, you will
need to set up a Dropbox App within your Dropbox account. It is
important that certain Dropbox App settings are edited before running
\code{activate_dropbox()}, including:
\itemize{
\item Under "OAUTH2: Redirect URIs" add the URI "http://localhost:1410/"
\item Under "Permissions" tick for access to Dropbox folder to read and write files.
}

Full details on completing these changes are detailed in Vignette 1 within the
BIOWELL package. Then you can generate your Dropbox App's Key and Secret in
the settings tab.

Once you have created a Dropbox App, the details (name, Key and Secret) can be
plugged into \code{activate_dropbox()}, alongside the path to your BIO-WELL survey
Shiny App folder (argument \code{BW_app_path}).

This function will generate a long-lived, refreshable Dropbox access token and
store it within a .secrets folder within your BIO-WELL survey Shiny App
folder. This enables the token to be accessible whilst hosted on the Shiny
Server.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

BW_app_path <- paste0(tempdir(),"/my_app_name_1")

dir.create(BW_app_path) # Create App directory to store token

activate_dropbox(Dropbox_App_Name = "my_app_name_1",
                Dropbox_App_Key = "your_app_key_here",
                Dropbox_App_Secret = "your_app_secret_here",
                BW_app_path = BW_app_path)
\dontshow{\}) # examplesIf}
}
