% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_URL.R
\name{create_URL}
\alias{create_URL}
\title{Create shiny survey with BIOWELL scale}
\usage{
create_URL(
  BW_app_path,
  appName,
  shiny_server_username,
  shiny_server_token,
  shiny_server_secret
)
}
\arguments{
\item{dates}{a}
}
\value{
"dd"
}
\description{
Function to generate custom shiny survey with BIOWELL scale
}
\details{
d
}
