% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Combine survey responses into data frame}
\usage{
download_data(drop_box_folder, BW_app_path)
}
\arguments{
\item{drop_box_folder}{path to folder in Dropbox. Defaults to the root directory.}
}
\value{
Data frame containing combined responses to survey
}
\description{
Function accesses survey responses from Dropbox folder and combines into single data frame.
}
\details{
With multiple survey responses stored in a Dropbox cloud storage folder, this
function downloads all responses and combined into a single data frame for
analysis of BIO-WELL scores across participants.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

data<-extract_biowell("responses")
\dontshow{\}) # examplesIf}
}
