% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{explore_data}
\alias{explore_data}
\title{Process and plot BIO-WELL scores}
\usage{
explore_data(
  data,
  column_name,
  var_name,
  biowell_scale,
  plot = TRUE,
  score = "inverted"
)
}
\arguments{
\item{data}{a data frame containing the combined response data frames from a
\code{BIOWELL} package survey. See \code{download_data()} to obtain this.}

\item{column_name}{optional; a character string, the name of the column in
\code{data} that contains categorical responses to split BIO-WELL scores by
before plotting.}

\item{var_name}{optional; the name for the variable to appear on the y axis of
plot.}

\item{biowell_scale}{optional; a character string, a single wellbeing domain
to plot instead of full BIO-WELL scores. One of; \code{physical}, \code{emotional},
\code{cognitive}, \code{social} and \code{spiritual.} See details for more information.}

\item{plot}{optional; a logical, indicating whether to show plots in the R
window. Default is \code{TRUE.}}

\item{score}{a character string, specifying whether to plot inverted or raw
BIO-WELL scores. One of; \code{inverted} or \code{raw} Default; \code{inverted}. See
details for more information.}
}
\value{
a list object containing the plots and processed data frame.
}
\description{
Function to generate plots of BIO-WELL scores across questions and groups from
BIO-WELL survey response data.
}
\details{
This function can be used to plot average BIO-WELL scores from participant
responses to your custom BIO-well survey (generated using \code{build_survey()}).
Please see function \code{download_data()} for extracting your BIO-WELL survey
response data from Dropbox.

All plots are created using the R package \code{ggplot2} (Wickham and Chang, 2016).

There are two approaches:
\itemize{
\item Default: This will generate a box plot of BIO-WELL scores for each BIO-WELL
question (\code{biowell_questions} in \code{build_survey()}) within each BIO-WELL
environmental space setting (\code{biowell_setting} in \code{build_surveys()}). If more than one
environmental space setting exists within your survey, these will be plotted in seperately.
\item Column: If a \code{column_name} is provided, for each unique category in this
column, a box plot of average BIO-WELL scores for each participant will be
generated. This can be used to compare average BIO-WELL scores across
responses to other questions with your BIO-WELL survey (e.g. age categories or
Likert scale responses)
}
}
\section{BIO-WELL scores}{
By default, the typical, inverted BIO-well score is plotted. This is
calculated by taking the average score from across five wellbeing domains
(physical, emotional, cognitive, social and spiritual). When inverted, the
BIO-WELL score can be interpreted as 0-49, a negative wellbeing response to
biodiversity, 50, a neutral wellbeing responses and 51-100 representing a
positive wellbeing response to biodiversity.The \code{raw} BIO-WELL scores
represents the opposite of this with 0-49, a positive wellbeing response to
biodiversity, 50, a neutral wellbeing responses and 51-100 representing a
negative wellbeing response.

Please see Irvine et al., (2023) for more details on the BIO-WELL score.

However, if you would like to only look at one of the wellbeing domains,
then you can use the argument \code{biowell_scale} to specify which to plot.

Please see Vignette 4 in the \code{BIOWELL} package for full guidance on this
function.
}

\examples{

data("sample_BW_data")

# Average BIO-WELL scores for each question across all participants
results <- explore_data(sample_BW_data)

# Average BIO-WELL score for each Gender response category
results <- explore_data(sample_BW_data,
                       column_name = "Gender")

# Average score in the emotional wellbeing domain of BIO-WELL across age groups.
results <- explore_data(sample_BW_data,
                       column_name = "Age",
                       biowell_scale = "emotional")

# Average BIO-WELL scores across sight impairment responses
results <- explore_data(sample_BW_data,
                       column_name = "impairment.sight")

}
\references{
Irvine, K.N., Fisher, J.C., Bentley, P.R., Nawrath, M., Dallimer,
M., Austen, G.E., Fish, R. and Davies, Z.G., 2023. BIO-WELL: The development
and validation of a human wellbeing scale that measures responses to
biodiversity. Journal of Environmental Psychology, 85, p.101921.doi:10.1016/j.jenvp.2022.101921.

Wickham, H. and Chang, W. , 2016. Package ‘ggplot2’. Create
elegant data visualisations using the grammar of graphics. Version, 2(1),
pp.1-189.
}
