% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_BW_data}
\alias{sample_BW_data}
\title{Sample BIO-WELL survey response data}
\format{
A data frame with 300 rows and 191 variables:
\describe{
\item{survey_duration}{The length of time participants took answering survey questions.}
\item{start_date}{The date that participants started the survey (format: YYYY-MM-DD).}
\item{start_time}{The time when participants started the survey (fprmat: HH:MM:SS)}
\item{end_date}{The date that participants completed the survey (format: YYYY-MM-DD).}
\item{end_time}{The time when participants finished the survey (fprmat: HH:MM:SS)}
\item{timezone}{The time zone in which start_time and end_time is recorded.}
\item{timezone_location}{The geographic location of the time zone.}
\item{server_submit_time}{The time when participants finished the survey in the server's time zone.}
\item{Consent}{Participants responses to the survey consent question.}
\item{Age}{Participants responses to the survey consent question.}
\item{Gender}{The age range of participants.}
\item{Country}{The country in which participants reside.}
\item{Urbanicity}{Whether participants considered themselves to be living in a rural or urban area.}
\item{forest_Encountering.living.things_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Encountering.living.things_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Number.of.living.things_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.living.things_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Interactions.between.living.things_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Living.processes_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.sounds_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Distinctive.sounds_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.colours_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Vivid.colours_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.shapes_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Maturity.of.living.things_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.textures_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Sponginess.of.living.things_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Variety.of.smells_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Woody.smells_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Changes.in.this.season_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Physical_RAW}{The raw score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Physical_INVERTED}{The inverted score from the physical wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Emotional_RAW}{The raw score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Emotional_INVERTED}{The inverted score from the emotional wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Cognitive_RAW}{The raw score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Cognitive_INVERTED}{The inverted score from the cognitive wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Social_RAW}{The raw score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Social_INVERTED}{The inverted score from the social wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Spiritual_RAW}{The raw score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{forest_Presence.of.animals_Spiritual_INVERTED}{The inverted score from the spiritual wellbeing domain slider scale (component of the BIO-WELL scale) in response to the question 'encountering the living things (e.g. plants, fungi and animals) in this woodland makes me feel…'})
\item{visits_aschild}{The extent to which participants agreed that they spent a lot of time in woodlands as a child.}
\item{visits_asteen}{The extent to which participants agreed that they spent a lot of time in woodlands as a teen.}
\item{visits_asadult}{The extent to which participants agreed that they spent a lot of time in woodlands as a adult.}
\item{impairment.smell}{Whether participants had a smell impairment or not.}
\item{impairment.sound}{Whether participants had a sound impairment or not.}
\item{impairment.sight}{Whether participants had a sight impairment or not.}
\item{mean_biowell_INVERTED}{The average BIO-WELL score across all bio-well questions for each participant (inverted; 100-raw).}
\item{mean_biowell_RAW}{The average BIO-WELL score across all bio-well questions for each participant (raw)}
}
}
\usage{
sample_BW_data
}
\description{
A dataset containing a sample of participant responses to a
BIO-WELL survey aim to relate woodland biodiversity to human wellbeing.
These data were collected for case study five from the study by Irvine et
al., (2023).
}
\references{
Irvine, K.N., Fisher, J.C., Bentley, P.R., Nawrath, M., Dallimer,
M., Austen, G.E., Fish, R. and Davies, Z.G., 2023. BIO-WELL: The development
and validation of a human wellbeing scale that measures responses to
biodiversity. Journal of Environmental Psychology, 85, p.101921.
}
\keyword{datasets}
