% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOWELL-package.R
\docType{package}
\name{BIOWELL-package}
\alias{BIOWELL}
\alias{BIOWELL-package}
\title{BIOWELL: Create shiny surveys with the BIO-WELL scale}
\description{
The BIO-WELL scale is a valid and relaible scale with strong psychometric properties that can be used to measure the human wellbeing responses to biodiversity. This package contains functions for developing and deploying customised shiny surveys that utilise the BIO-WELL scale.
}
\author{
\strong{Maintainer}: Rachel Dobson \email{eerdo@leeds.ac.uk} (\href{https://orcid.org/0000-0003-3990-267X}{ORCID}) [contributor]

Authors:
\itemize{
  \item Katherine N. Irvine \email{kate.irvine@hutton.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
  \item Jessica C. Fisher \email{J.C.Fisher@kent.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
  \item Phoebe R. Bentley \email{xxxx} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
  \item Maximillian Nawrath \email{eemn@leeds.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
  \item Martin Dallimer \email{M.Dallimer@leeds.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
  \item Gail E. Austen \email{G.E.Austen@kent.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
  \item Rob Fish \email{r.fish@kent.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
  \item Zoe G. Davies \email{Z.G.Davies@kent.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
}

}
\keyword{internal}
