% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_survey.R
\name{build_survey}
\alias{build_survey}
\title{Create BIO-WELL survey in Shiny}
\usage{
build_survey(
  survey_title = NULL,
  sidepanel_message = NULL,
  Dropbox_App_folder,
  BW_app_path,
  organisation,
  organisation_website = NULL,
  screen_message = NULL,
  screen_questions,
  screen_questions_ID,
  screen_response_options = NULL,
  start_message = NULL,
  start_questions,
  start_questions_ID,
  start_questions_type,
  start_response_options = NULL,
  biowell_situations,
  biowell_situations_ID,
  biowell_questions,
  biowell_questions_ID,
  end_message = NULL,
  end_questions,
  end_questions_type,
  end_questions_ID,
  end_response_options = NULL,
  all_questions = TRUE,
  all_sliders = TRUE,
  user_report = TRUE,
  language = "english",
  offline_mode = FALSE,
  return = "run"
)
}
\arguments{
\item{survey_title}{a character string, the title of the survey to display on
each page.}

\item{sidepanel_message}{a character string, the text to display on the side
panel throughout the survey.}

\item{Dropbox_App_folder}{a character string, the path to the folder in your
Dropbox App to save survey responses to.}

\item{BW_app_path}{a character string, the path to your BIO-WELL survey Shiny
App folder. See details for more information.}

\item{organisation}{a character string, the name of the organisation running
the survey.}

\item{organisation_website}{a character string, the web address for the
organisation running the survey.}

\item{screen_message}{a character string, the text to display above the
screening questions.}

\item{screen_questions}{a character string or vector, the text for each
screening question.}

\item{screen_questions_ID}{a character string or vector equal to the length of
\code{screen_questions}, providing short IDs for each screening question to use
in output survey response data frame.}

\item{screen_response_options}{a list the length of \code{screen_questions} with
each element containing a vector of choices for responding to each screening
question in-turn. To specify which answers to screen by add ":SCREEN" to the
option. See details for an example.}

\item{start_message}{a character string, the text to display above the "start"
questions.}

\item{start_questions}{a character string or vector, the "start" questions to
ask participants prior to the BIO-WELL questions.}

\item{start_questions_ID}{a character string or vector equal to the length of
\code{start_questions}, providing short IDs for each "start" question to use in
output survey response data frame.}

\item{start_questions_type}{a character string or vector, the type
of response item for each "start" question. Length must be equal to the
length of \code{start_questions}. One of; \code{text}, \code{checkbox}, \code{selectbox},
\code{likert_five} and \code{likert_seven}. See details for more information.}

\item{start_response_options}{a list of character strings or
vectors, the options to offer for each \code{start_questions} of type \code{checkbox}
or \code{selectbox}. Provide these in the same order as they appear in
\code{start_questions}.}

\item{biowell_situations}{a character string or vector, describing the
environmental space setting(s) for BIO-WELL questions.}

\item{biowell_situations_ID}{a character string or vector the length of
\code{biowell_situations}, the short IDs for each BIO-WELL situation to use in
output survey response data frame.}

\item{biowell_questions}{a list of character strings or vectors, the questions
for each separate \code{biowell_situations}. Each will be accompanied by the
BIO-WELL response item comprising five sliders.}

\item{biowell_questions_ID}{a character string or vector equal to the length
of \code{biowell_questions}, the short IDs for each BIO-WELL question to use in
output survey response data frame.}

\item{end_message}{a character string, the text to display above the "end"
questions.}

\item{end_questions}{a character string or vector, the questions to
ask participants following the BIO-WELL questions.}

\item{end_questions_type}{a character string or vector, the type of
response item for each "end" question. Length must be equal to the length of
\code{start_questions}. One of; \code{text}, \code{checkbox}, \code{selectbox}, \code{likert_five}
and \code{likert_seven}. See details for more information.}

\item{end_questions_ID}{a character string or vector equal to the length of
\code{end_questions}, providing short IDs for each "end" question to use in
output survey response data frame.}

\item{end_response_options}{a list of character strings or vectors,
the options to offer for each \code{end_questions} of type \code{checkbox} or
\code{selectbox}. Provide these in the same order as they appear in
\code{end_questions}.}

\item{all_questions}{a logical, indicating whether all "start" and "end"
questions must be answered to continue with the survey.}

\item{all_sliders}{a logical, indicating whether all BIO-WELL sliders must be
moved at least once to continue with the survey.}

\item{user_report}{a logical, indicating whether to generate a BIO-WELL score
report for participants. See details for more information.}

\item{language}{a character string, the language for the BIO-WELL sliders. See
details for all 31 language options available. Default; \code{english}.}

\item{offline_mode}{Optional; a logical, indicating whether to upload survey
response data to Dropbox. Useful when developing the survey.}

\item{return}{optional; whether to run the survey Shiny app or return app
component. One of; \code{run}, \code{ui}, \code{server.} Default: \code{run}.}
}
\value{
Generates a custom BIO-WELL survey Shiny App.
}
\description{
Function to generate custom BIO-WELL survey with R Shiny
}
\section{BIO-WELL survey structure}{
\itemize{
\item Screening: This section allows you to set questions where certain responses
will prevent users from completing the survey. For instance, if they are not
of a certain age, or do not consent their involvement.
\item Start: This section allows you to set questions that come before the BIO-WELL
questions. For example, these may be generic questions, such as gaining
information on a participant's location, gender or employment.
\item BIO-WELL: The questions in this section are all accompanied by the five
sliders that comprise the BIO-WELL scale. For each situation given (although
often there may be only be one environment of interest), a new page is added
to the survey. See Irvine et al., (2023) for more details on the BIO-WELL
scale and survey set-up, including the BIO-WELL stem questions.
}
\enumerate{
\item \code{biowell_situations}: Text for describing an introductory scenario to
cognitively situate the participant within a particular natural environment
(i.e. a nearby forest).
\item \code{biowell_questions}: A list the length of \code{biowell_situations}, in which
each element contains a vector of the questions to be asked for each
situation. These stem questions are asked about biodiversity metrics (e.g.
abundance, species diversity), as well as biodiversity attributes (e.g.
smells, colours, shapes) within each \code{biowell_situation} given.
}
\itemize{
\item End: This section allows you to set questions that follow the BIO-WELL
questions. This may include asking participants for their contact details or
questions that may help disentangle patterns in BIO-WELL scores, such as
whether individuals spend a lot of time outdoors.
\item Report: This section will calculate the participants average BIO-WELL score
across the five wellbeing domains in the BIO-WELL scale (physical, emotional,
cognitive, social and spiritual) and provide users with an interpretation of
their BIO-WELL score, including a comparison with other survey participants.
}
}

\section{Structure: messages, question, IDs, types and options.}{
Each section can be specified using arguments relating to the following
aspects:
\subsection{Messages}{

Text that will accompany each section or display throughout the survey.
}

\subsection{Questions and IDs}{

Questions provide the exact text for each question to ask participants.

IDs provide shorted versions or codes for each question within the survey.
These are then used in the output survey response data frame to prevent
lengthy column names. For example, the question "The variety of textures in
this forest makes me feel…" could be given the ID "Variety of textures".
}

\subsection{Response types and options}{

There are five response type options to accompany non-BIO-WELL questions.
These include:
\enumerate{
\item \code{text} - A simple text box that the participant can type their answer
into.
\item \code{selectbox}* - A drop down box that the participant can choose one
answer from.
\item \code{checkbox}* - A group of check boxes that the participant can choose multiple
or none of. Be aware that if \code{all_questions = TRUE}, participants will have to
select one of the check boxes, and so it is best to include an option for
selecting none of the options below.
\item \code{likert_five} - A five point Likert scale going from "strongly
agree" to "strongly disagree".
\item \code{likert_seven} - A seven point Likert scale that also includes
"somewhat agree" and "somewhat disagree".
}

*For \code{selectbox} and \code{checkbox} response types, you will need to set the options
that participants will have. This can be done using the "...response_options"
arguments.

To specify this, for each \code{selectbox} and \code{checkbox} response type in each
section, create a list of equal length to the number of these response types.
Each element within the list should contain a character vector of the options.

For example:

start_questions <- c("How old are you?",
"What is your current employment?",
"Which of the following pets do you own?")

start_questions_ID <- c("Age","Employment","Pets")

start_questions_type <- c("selectbox", "text", "checkbox")

start_response_options <- list(c("0-17", "18-49", "51 +"),
c("dog", "cat", "rabbit", "fish", "snake"))
}
}

\section{Screening questions}{
Screening questions offer you the opportunity to remove participants from
taking part in the survey if they do not meet certain criteria. These will all
be \code{selectbox} response types. To specify how to screen participants, add
":SCREEN" to the response options for which, if chosen, the participant should
be removed.

For example, if participants do not consent or are under 18, they will be
removed from the following survey:

screen_questions <- c("Do you consent to this study?","How old are you?")

screen_questions_ID <- c("Consent","Age")

screen_response_options <- list(c("No:SCREEN", "Yes"),
c("0-17:SCREEN", "18-49", "51+"))
}

\section{Happy with your BIO-WELL survey design?}{
Once you have customised your survey, you need to save the R script as "app.R"
within a designated folder for your BIOWELL survey Shiny App.

This R script should contain nothing other than "BIOWELL::build_survey()" with
the arguments to the function in-filled between the brackets. See Vignette 3
for more details.

You can then deploy your survey onto the Shiny Server using the \code{create_URL()}
function.
}

\section{Language options}{
The BIO-WELL response sliders can be generated in 31 different languages. The
language for your study can be specified using the \code{language} argument. The
available options are: + \code{arabic} + \code{bengali_india} + \code{chinese_hong_kong} +
\code{chinese_taiwan} + \code{danish} + \code{dutch} + \code{edo} + \code{english} + \code{finnish} +
\code{french} + \code{german} + \code{gujarati} + \code{indonesian} + \code{irish} + \code{italian} +
\code{japanese} + \code{korean} + \code{malay} + \code{norwegian} + \code{polish} + \code{portuguese} +
\code{portuguese_brasil} + \code{punjabi_india} + \code{punjabi_pakistan} + \code{russian} +
\code{scottish_gaelic} + \code{spanish} + \code{swedish} + \code{urdu_india} + \code{urdu_pakistan} +
\code{welsh}.

Please see Vignette 2 in the \code{BIOWELL} package for full guidance on this
function.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
build_survey(
survey_title = "Insert survey title here",
 sidepanel_message = "Insert sidepanel message here.",
 start_message = "Insert start message here",
 screen_message = "Insert screen message here",
 screen_questions = c("Insert participant screening question here"),
 screen_questions_ID = c("Screening_Q1"),
 screen_response_options = list(c(
   "You will be removed:SCREEN",
   "You will not be removed"
 )),
 start_questions = c(
   "Insert start question: text",
   "Insert start question: selectbox",
   "Insert start question: checkbox",
   "Insert start question: likert_five",
   "Insert start question: likert_seven"
 ),
 start_questions_ID = c(
   "text_SQ1",
   "selectbox_SQ2",
   "checkbox_SQ3",
   "likert_five_SQ4",
   "likert_seven_SQ5"
 ),
 start_questions_type = c("text",
                          "selectbox",
                          "checkbox",
                          "likert_five",
                          "likert_seven"),
 start_response_options = list(
   c('Choose',  'one',  'of', 'these'),
   c('Choose',  'multiple',  'of', 'these')
 ),
biowell_situations = c(
  "Insert description of biowell situation 1.",
   "Insert description of biowell situation 2."
 ),
biowell_situations_ID = c("situation1", "situation2"),
 biowell_questions = list(
   c("Insert biodiversity stem question here BW1"),
   c(
     "Insert biodiversity stem question here - BW2",
     "Insert biodiversity stem question here - BW3"
   )
 ),
 biowell_questions_ID = list(c("BW.S1.Q1"),
                             c("BW.S2.Q1",
                               "BW.S2.Q2")),
 end_questions = c(
   "Insert end question: text",
   "Insert end question: selectbox",
   "Insert end question: checkbox",
  "Insert end question: likert_five",
   "Insert end question: likert_seven"
 ),
 end_questions_ID = c(
   "text_EQ1",
   "selectbox_EQ2",
   "checkbox_EQ3",
   "likert_five_EQ4",
   "likert_seven_EQ5"
 ),
 end_questions_type = c("text",
                        "selectbox",
                        "checkbox",
                        "likert_five",
                        "likert_seven"),
 end_response_options = list(
   c('Choose',  'one',  'of', 'these'),
  c('Choose',  'multiple',  'of', 'these')
),
 end_message = "Insert end message here.",
 Dropbox_App_folder = "insert_Dropbox_App_folder_name",
 BW_app_path = tempdir(),
 organisation = "BIOWELL package example",
 organisation_website = "https://github.com/r-a-dobson/BIOWELL",
 all_questions = TRUE,
 all_sliders = FALSE,
 user_report = TRUE,
 offline_mode = TRUE
)
\dontshow{\}) # examplesIf}
}
\references{
Irvine, K.N., Fisher, J.C., Bentley, P.R., Nawrath, M., Dallimer,
M., Austen, G.E., Fish, R. and Davies, Z.G., 2023. BIO-WELL: The development
and validation of a human wellbeing scale that measures responses to
biodiversity. Journal of Environmental Psychology, 85, p.101921.
}
