% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_URL.R
\name{create_URL}
\alias{create_URL}
\title{Deploy BIO-WELL survey to Shiny Server}
\usage{
create_URL(
  BW_app_path,
  BW_app_name,
  Shiny_Server_user,
  Shiny_Server_token,
  Shiny_Server_secret
)
}
\arguments{
\item{BW_app_path}{a character string, the path to your BIO-WELL Shiny App
folder.}

\item{BW_app_name}{a character string, a name for your BIO-WELL Shiny App.}

\item{Shiny_Server_user}{a character string, your Posit Shiny Server username.}

\item{Shiny_Server_token}{a character string, your Posit Shiny Server Token.}

\item{Shiny_Server_secret}{a character string, your Posit Shiny Server Secret.}
}
\value{
a character string, the URL to your deployed BIO-WELL survey Shiny
App.
}
\description{
Function to deploy BIO-WELL survey Shiny App to the Shiny Server and return a
URL for survey distribution to participants.
}
\details{
To share the BIO-WELL survey that you have built using \code{build_survey()} with
participants, we provide \code{create_URL()} to upload your BIOWELL survey Shiny
App to the \href{https://posit.co/products/open-source/shinyserver/}{Posit Shiny Server}.

For this, you will first need to register for a free Posit Shiny Server
account on their website. This will provide you with your Shiny Server
username, token and secret for input into the \code{create_URL()} function, which
acts as a wrapper for the \code{rsconnect} \href{https://CRAN.R-project.org/package=rsconnect}{R package}.

To deploy the app successfully, ensure that you have:
\enumerate{
\item Activated a long-lived Dropbox access token. This should be stored within
your BIO-WELL survey Shiny App (automatically done using \code{activate_dropbox()}
function).
\item Within your BIO-WELL survey Shiny App folder, an R script named "app.R"
should be saved. This R script should contain your input into
\code{build_survey()}. Please remember to include nothing in the script apart from
this function and your filled arguments (e.g. no "library(BIOWELL)" or
"setwd()"), and to specify the package that the \code{build_survey()} function is
from by using the double-colon operator (i.e. \code{BIOWELL::build_survey()} with
your arguments within the brackets).
}

Once the function has been run, the BIO-WELL survey Shiny App will be deployed
to your Shiny Server account. This may take a couple of minutes, and will
finish by loading the URL in your browser and returning the URL in the R
Console.

If you are re-uploading the BIO-WELL survey for any reason, look out for a
message to confirm this by running 'Y' in the R Console.

The URL can now be shared to participants for online access to your BIO-WELL
survey. Responses will be remotely saved to your Dropbox account.You can check
the status of your App on the posit Shiny Server app dashboard.

Please see Vignette 3 in the \code{BIOWELL} package for full guidance on this
function.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

BW_app_path <- paste0(tempdir(),"/my_app_name_1")

dir.create(BW_app_path) # Store your "app.R" script here

BIOWELL::create_URL(BW_app_path = paste0(tempdir(),"/my_app_name_1"),
                   BW_app_name = "my_app_name_1",
                   Shiny_Server_user = 'your_username_here',
                   Shiny_Server_token = 'your_token_here',
                   Shiny_Server_secret = 'your_secret_here')

\dontshow{\}) # examplesIf}
}
