% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_survey.R
\name{evaluate_survey}
\alias{evaluate_survey}
\title{Evaluate BIO-WELL survey}
\usage{
evaluate_survey(data)
}
\arguments{
\item{data}{a data frame containing the combined response data frames from a
\code{BIOWELL} package survey. See \code{download_data()} to obtain this.}
}
\value{
Returns a list containing a data frame of processed BIO-WELL scores,
Cronbach's alpha statistic and item-total.
}
\description{
Function to calculate key survey evaluation statistics on BIO-WELL survey
responses.
}
\details{
Once you have combined all responses to your BIO-WELL survey into a single
data frame using the \code{download_data()} function, this data frame can be input
into \code{evaluate_survey()}.

\code{evaluate_survey()} returns three outputs in a list object:
\enumerate{
\item A processed data frame containing the average BIO-WELL scores for each
participant for each questions within your custom \code{BIOWELL} package survey.
\item Cronbach's Alpha: the computed Cronbach's Alpha statistic for your survey,
which measures the internal consistency of responses. This is calculated using
the \code{cronbach.alpha()} function from the \code{ltm} R package (Rizopoulos et al.,
2007).
\item Item-total correlation: the computed item-total statistic for each BIO-WELL
question within your survey (Cronbach et al., 1951). This is calculated using
the \code{item.total()} function from the \code{multilevel} \href{https://CRAN.R-project.org/package=multilevel}{R package}.
}
}
\examples{
# Read in sample BIO-WELL survey data
data("sample_BW_data")

results<-evaluate_survey(sample_BW_data)

# Data frame containing average BIO-WELL scores for each participant x
# question

head(results[[1]])

# Cronbach's alpha
results[[2]]

# Item total
results[[3]]
}
\references{
Cronbach, L. J. (1951) Coefficient alpha and the internal structure of tests.
Psychometrika, 16, 297–334.

Rizopoulos, D., 2007. ltm: An R package for latent variable modeling and item
response analysis. Journal of statistical software, 17, pp.1-25.
}
