% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_survey.R
\name{evaluate_survey}
\alias{evaluate_survey}
\title{Create shiny survey with BIOWELL scale}
\usage{
evaluate_survey(data, type = "situation_question")
}
\arguments{
\item{survey_title}{a character string, the title of the survey to display on each page.}

\item{biowell_situations}{a character string or vector, describing the environmental space setting for BIO-WELL questions.}

\item{biowell_questions}{a list of character strings or vectors, the questions for each separate \code{biowell_situations} for BIO-WELL responses.}

\item{start_message}{a character string, a message to display to participants at the start of the survey.}

\item{start_questions}{optional; a character string or vector, the questions to ask participants at the start of the survey.}

\item{start_questions_type}{optional; a character string or vector, the type of response for each start question. Must be length of \code{start_questions}. One of; \code{text}, \code{checkbox} or \code{selectbox}. See details for information.}

\item{start_drop_down_options}{optional; a list of character strings or vectors, the options to offer for each \code{start_questions} of type \code{checkbox} or \code{selectbox}.}

\item{end_message}{a character string, a message to display to participants at the end of the survey.}

\item{end_questions}{optional; a character string or vector, the questions to ask participants at the end of the survey.}

\item{end_questions_type}{optional; a character string or vector, the type of response for each end question. Must be length of \code{end_questions}. One of; \code{text}, \code{checkbox} or \code{selectbox}. See details for information.}

\item{end_drop_down_options}{optional; a list of character strings or vectors, the options to offer for each \code{end_questions} of type \code{checkbox} or \code{selectbox}.}

\item{organisation_website}{a character string, the web address for the organisation running the survey.}

\item{organisation}{a character string, the name of the organisation running the survey.}

\item{drop_box_folder}{path to folder in Dropbox to save responses to. Defaults to the root directory.}
}
\value{
runs a customised shiny survey app or returns UI / server components for a shiny survey.
}
\description{
Function to generate custom shiny survey with BIOWELL scale
}
\details{
df
}
\examples{
results<-evaluate_survey(data = example_survey_data, type = "situation_question")
results<-evaluate_survey(data = example_survey_data, type = "question")

}
